+++ library/core/src/main/java/com/google/android/exoplayer2/source/DeferredMediaPeriod.java
@@ -15,6 +15,7 @@
  */
 package com.google.android.exoplayer2.source;
 
+import android.support.annotation.Nullable;
 import com.google.android.exoplayer2.SeekParameters;
 import com.google.android.exoplayer2.source.MediaSource.MediaPeriodId;
 import com.google.android.exoplayer2.trackselection.TrackSelection;
@@ -29,6 +30,15 @@ import java.io.IOException;
  */
 public final class DeferredMediaPeriod implements MediaPeriod, MediaPeriod.Callback {
 
+  /** Listener for preparation errors. */
+  public interface PrepareErrorListener {
+
+    /**
+     * Called the first time an error occurs while refreshing source info or preparing the period.
+     */
+    void onPrepareError(IOException exception);
+  }
+
   public final MediaSource mediaSource;
 
   private final MediaPeriodId id;
@@ -37,11 +47,31 @@ public final class DeferredMediaPeriod implements MediaPeriod, MediaPeriod.Callb
   private MediaPeriod mediaPeriod;
   private Callback callback;
   private long preparePositionUs;
+  private @Nullable PrepareErrorListener listener;
+  private boolean notifiedPrepareError;
 
+  /**
+   * Creates a new deferred media period.
+   *
+   * @param mediaSource The media source to wrap.
+   * @param id The identifier for the media period to create when {@link #createPeriod()} is called.
+   * @param allocator The allocator used to create the media period.
+   */
   public DeferredMediaPeriod(MediaSource mediaSource, MediaPeriodId id, Allocator allocator) {
     this.id = id;
     this.allocator = allocator;
     this.mediaSource = mediaSource;
+  }
+
+  /**
+   * Sets a listener for preparation errors.
+   *
+   * @param listener An listener to be notified of media period preparation errors. If a listener is
+   *     set, {@link #maybeThrowPrepareError()} will not throw but will instead pass the first
+   *     preparation error (if any) to the listener.
+   */
+  public void setPrepareErrorListener(PrepareErrorListener listener) {
+    this.listener = listener;
   }
 
   /**
@@ -76,10 +106,20 @@ public final class DeferredMediaPeriod implements MediaPeriod, MediaPeriod.Callb
 
   @Override
   public void maybeThrowPrepareError() throws IOException {
-    if (mediaPeriod != null) {
-      mediaPeriod.maybeThrowPrepareError();
-    } else {
-      mediaSource.maybeThrowSourceInfoRefreshError();
+    try {
+      if (mediaPeriod != null) {
+        mediaPeriod.maybeThrowPrepareError();
+      } else {
+        mediaSource.maybeThrowSourceInfoRefreshError();
+      }
+    } catch (final IOException e) {
+      if (listener == null) {
+        throw e;
+      }
+      if (!notifiedPrepareError) {
+        notifiedPrepareError = true;
+        listener.onPrepareError(e);
+      }
     }
   }